last

+++ SwGohForms/Form1.cs
@@ -25,23 +25,41 @@ namespace SwGohForms
         private void butGuildLoad_Click(object sender, EventArgs e)
         {
             //ExecuteCommand(Command.UpdateUnknownGuild, "Order 66 501st Division#@#32#@#order-66-501st-division", null);
-            string pathToProgram = "..\\Forms/ConsoleSwGohParser.Core.dll";
+            string pathToProgram = Application.StartupPath + "\\..\\Parser\\ConsoleSwGohParser.Core.dll";
             if (File.Exists(pathToProgram))
             {
+                pathToProgram = "\"" + pathToProgram + "\"";
+
                 string guildname = textGuildName.Text;
                 string guildID = textGuildID.Text;
                 string guildfixedname = textGuildFixed.Text;
 
                 string argsString = pathToProgram + " " + "UpdateUnknownGuild" + " \"" + guildname + "#@#" + guildID + "#@#" + guildfixedname + "\"";
 
-                var process = Process.Start("dotnet", argsString);
-                process.WaitForExit();
-                var exitCode = process.ExitCode;
+                //var process = Process.Start("dotnet", argsString);
+                //process.WaitForExit();
+                //var exitCode = process.ExitCode;
+
+                var process = new Process
+                {
+                    StartInfo = new ProcessStartInfo
+                    {
+                        FileName = "dotnet",
+                        Arguments = argsString,
+                        UseShellExecute = true,
+                        RedirectStandardOutput = false,
+                        RedirectStandardError = false,
+                        CreateNoWindow = true
+                    }
+
+                };
+
+                process.Start();
             }
-            else
-            {
-                MessageBox.Show("Cannot Find " + pathToProgram);
-            }
+            //else
+            //{
+            //    MessageBox.Show("Cannot Find " + pathToProgram);
+            //}
         }
 
         private void butDelFromQ_Click(object sender, EventArgs e)
@@ -59,16 +77,21 @@ namespace SwGohForms
                     if (collection != null)
                     {
                         string guildname = textDelFromQ.Text;
-                        FilterDefinition<QueueDto> filter2 = Builders<QueueDto>.Filter.Eq("Guild", guildname);
-                        DeleteResult res2 = collection.DeleteMany(filter2);
 
-                        if (res2.IsAcknowledged)
+                        DialogResult ress = MessageBox.Show(this, "Deleting from Queue for guild : " + guildname, "Caution", MessageBoxButtons.YesNo, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button1);
+                        if (ress == DialogResult.Yes)
                         {
-                            MessageBox.Show(res2.DeletedCount.ToString () + " Queues Deleted");
-                        }
-                        else
-                        {
-                            MessageBox.Show("Result not Acknowledged");
+                            FilterDefinition<QueueDto> filter2 = Builders<QueueDto>.Filter.Eq("Guild", guildname);
+                            DeleteResult res2 = collection.DeleteMany(filter2);
+
+                            if (res2.IsAcknowledged)
+                            {
+                                MessageBox.Show(res2.DeletedCount.ToString() + " Queues Deleted");
+                            }
+                            else
+                            {
+                                MessageBox.Show("Result not Acknowledged");
+                            }
                         }
                     }
                 }
@@ -97,17 +120,22 @@ namespace SwGohForms
                     IMongoCollection<PlayerDto> collection = db.GetCollection<PlayerDto>("Player");
                     if (collection != null)
                     {
-                        string guildname = textDelFromQ.Text;
-                        FilterDefinition<PlayerDto> filter2 = Builders<PlayerDto>.Filter.Eq("GuildName", guildname);
-                        DeleteResult res2 = collection.DeleteMany(filter2);
+                        string guildname = textDelFromP.Text;
 
-                        if (res2.IsAcknowledged)
+                        DialogResult ress = MessageBox.Show(this, "Deleting from Player for guild : " + guildname, "Caution", MessageBoxButtons.YesNo, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button1);
+                        if (ress == DialogResult.Yes)
                         {
-                            MessageBox.Show(res2.DeletedCount.ToString() + " Players Deleted");
-                        }
-                        else
-                        {
-                            MessageBox.Show("Result not Acknowledged");
+                            FilterDefinition<PlayerDto> filter2 = Builders<PlayerDto>.Filter.Eq("GuildName", guildname);
+                            DeleteResult res2 = collection.DeleteMany(filter2);
+
+                            if (res2.IsAcknowledged)
+                            {
+                                MessageBox.Show(res2.DeletedCount.ToString() + " Players Deleted");
+                            }
+                            else
+                            {
+                                MessageBox.Show("Result not Acknowledged");
+                            }
                         }
                     }
                 }
